+++ src/Core/Query/FieldSets.php
@@ -21,6 +21,7 @@ namespace LaravelJsonApi\Core\Query;
 
 use Countable;
 use Illuminate\Contracts\Support\Arrayable;
+use Illuminate\Support\Collection;
 use Illuminate\Support\Enumerable;
 use IteratorAggregate;
 use UnexpectedValueException;
@@ -115,12 +116,12 @@ class FieldSets implements Arrayable, IteratorAggregate, Countable
     }
 
     /**
-     * @param string ...$keys
+     * @param string ...$resourceTypes
      * @return $this
      */
-    public function forget(string ...$keys): self
+    public function forget(string ...$resourceTypes): self
     {
-        foreach ($keys as $key) {
+        foreach ($resourceTypes as $key) {
             unset($this->stack[$key]);
         }
 
@@ -171,6 +172,14 @@ class FieldSets implements Arrayable, IteratorAggregate, Countable
     public function all(): array
     {
         return $this->stack;
+    }
+
+    /**
+     * @return Collection
+     */
+    public function collect(): Collection
+    {
+        return collect($this->stack);
     }
 
     /**