+++ src/Core/Query/IncludePaths.php
@@ -21,6 +21,7 @@ namespace LaravelJsonApi\Core\Query;
 
 use Countable;
 use Illuminate\Contracts\Support\Arrayable;
+use Illuminate\Support\Collection;
 use Illuminate\Support\Enumerable;
 use IteratorAggregate;
 use UnexpectedValueException;
@@ -170,6 +171,14 @@ class IncludePaths implements IteratorAggregate, Countable, Arrayable
     public function all(): array
     {
         return $this->stack;
+    }
+
+    /**
+     * @return Collection
+     */
+    public function collect(): Collection
+    {
+        return collect($this->stack);
     }
 
     /**